+++ AppLink/AppLink/FMCTCPTransport.m
@@ -111,7 +111,7 @@ static void TCPCallback(CFSocketRef socket, CFSocketCallBackType type, CFDataRef
 - (void) sendData:(NSData*) msgBytes {
 
 	NSString* byteStr = [self getHexString:msgBytes];
-    [FMCDebugTool logInfo:[NSString stringWithFormat:@"Sent %d bytes: %@", msgBytes.length, byteStr] withType:FMCDebugType_Transport_TCP toOutput:FMCDebugOutput_DeviceConsole];
+    [FMCDebugTool logInfo:[NSString stringWithFormat:@"Sent %lu bytes: %@", (unsigned long)msgBytes.length, byteStr] withType:FMCDebugType_Transport_TCP toOutput:FMCDebugOutput_DeviceConsole];
 
     CFSocketError e = CFSocketSendData(socket, NULL, (CFDataRef)msgBytes, 10000);
     if (e != kCFSocketSuccess) {
